implemented Cython on key A* functions : from 35sec to 18sec on complex Puzzle
