+++ Makefile
@@ -1,12 +1,18 @@
+PYTHON=python3
+
+build:
+	$(PYTHON) setup.py build_ext --inplace
+
 run:
 	python3 -m solver.main
 
 runfix:
 	python3 -m solver.main ./boards/npuzzle-3-1.txt
+
 test:
 	python3 -m unittest discover -s tests
 
 clean:
-	rm -rf solutions/*
+	rm -rf solutions/* build __pycache__ *.c *.so solver/*.c solver/*.so 
 
 .PHONY: run runfix test clean
